requirements-normdocs.txt
# Десктоп normdocs_app + сборка EXE (tkinter уже в Python)
python-dotenv>=1.0
requests>=2.31
pymupdf>=1.24
python-docx>=1.1
rarfile>=4.0
pyinstaller>=6.0
